clean up legacy config types
