+++ libs/adapters/src/rabbitmq/index.ts
@@ -1,4 +1,2 @@
 export * from './RabbitMQAdapter';
-export { createRmqConfig } from './createRmqConfig';
-export * from './types';
-export * from './util';
+export * from './createRmqConfig';
